backend/requirements.txt
Django
djangorestframework
django-cors-headers
channels
gunicorn
daphne
psycopg2-binary
whitenoise
python-dotenv
djangorestframework-simplejwt
django-anymail[mailgun]
